Roughly reorganize hooks into independent files (#27971)
