+++ client/web/src/enterprise/codeintel/indexes/hooks/queryCommitGraphMetadata.tsx
@@ -1,0 +1,56 @@
+import { ApolloClient } from '@apollo/client'
+import { parseISO } from 'date-fns'
+import { from, Observable } from 'rxjs'
+import { map } from 'rxjs/operators'
+
+import { gql, getDocumentNode } from '@sourcegraph/shared/src/graphql/graphql'
+
+import {
+    CodeIntelligenceCommitGraphMetadataResult,
+    CodeIntelligenceCommitGraphMetadataVariables,
+} from '../../../../graphql-operations'
+
+const GRAPH_METADATA = gql`
+    query CodeIntelligenceCommitGraphMetadata($repository: ID!) {
+        node(id: $repository) {
+            __typename
+            ... on Repository {
+                codeIntelligenceCommitGraph {
+                    stale
+                    updatedAt
+                }
+            }
+        }
+    }
+`
+
+export const queryCommitGraphMetadata = (
+    repository: string,
+    client: ApolloClient<object>
+): Observable<{ stale: boolean; updatedAt: Date | null }> =>
+    from(
+        client.query<CodeIntelligenceCommitGraphMetadataResult, CodeIntelligenceCommitGraphMetadataVariables>({
+            query: getDocumentNode(GRAPH_METADATA),
+            variables: { repository },
+        })
+    ).pipe(
+        map(({ data }) => data),
+        map(({ node }) => {
+            if (!node) {
+                throw new Error('Invalid repository')
+            }
+            if (node.__typename !== 'Repository') {
+                throw new Error(`The given ID is ${node.__typename}, not Repository`)
+            }
+            if (!node.codeIntelligenceCommitGraph) {
+                throw new Error('Missing code intelligence commit graph value')
+            }
+
+            return {
+                stale: node.codeIntelligenceCommitGraph.stale,
+                updatedAt: node.codeIntelligenceCommitGraph.updatedAt
+                    ? parseISO(node.codeIntelligenceCommitGraph.updatedAt)
+                    : null,
+            }
+        })
+    )
