+++ client/web/src/enterprise/codeintel/uploads/hooks/types.tsx
@@ -1,0 +1,54 @@
+import { gql } from '@sourcegraph/shared/src/graphql/graphql'
+
+export const lsifUploadFieldsFragment = gql`
+    fragment LsifUploadFields on LSIFUpload {
+        __typename
+        id
+        inputCommit
+        inputRoot
+        inputIndexer
+        projectRoot {
+            url
+            path
+            repository {
+                url
+                name
+            }
+            commit {
+                url
+                oid
+                abbreviatedOID
+            }
+        }
+        state
+        failure
+        isLatestForRepo
+        uploadedAt
+        startedAt
+        finishedAt
+        placeInQueue
+        associatedIndex {
+            id
+            state
+            queuedAt
+            startedAt
+            finishedAt
+            placeInQueue
+        }
+    }
+`
+
+export const lsifUploadConnectionFieldsFragment = gql`
+    fragment LsifUploadConnectionFields on LSIFUploadConnection {
+        nodes {
+            ...LsifUploadFields
+        }
+        totalCount
+        pageInfo {
+            endCursor
+            hasNextPage
+        }
+    }
+
+    ${lsifUploadFieldsFragment}
+`
